fidelity-csv is a full rewrite (Fidelity CSV ingestion, baskets, options
harness, quant suite, daily-advisory, and the daily catalyst news-bridge).
It supersedes main's earlier "Added Metrics to report" commit, whose
metrics (sharpe/calmar/drawdown) are already present in the rewrite's
reporting. Conflicts resolved in favor of the rewrite.

+++ src/trade_log/logger.py
@@ -32,7 +32,7 @@ class TradeLogger:
         self,
         ticker: str,
         action: str,
-        shares: int,
+        shares: float,
         price: float,
         rationale: str = "",
         tags: Optional[List[str]] = None,
